feat(hod|requisition): handle approval or rejection of requisition and their items
Requisitions and their corresponding items can be approved or rejected by hod. Similar operations can be executed on items in bulk or individually.
The items table accounts for checkbox validation based on items status with respect to check all.
A `hrReview` status was accounted for in the status badge.
Checkbox now supports `indeterminate` state.
Api requests has been modularised with `apiClient.ts`, corresponding feature service and constants for static strings.

Note: The bulk handle of requisition items is not fully functional as the api gives a 403 (forbidden) error.